proshop/requirements/base.txt
Django==4.2.10
psycopg2-binary==2.9.9
Pillow==10.0.0
djangorestframework==3.15.0
django-cors-headers==4.3.1
django-filter==24.1
django-crispy-forms==2.1
crispy-bootstrap5==2.0.2
python-decouple==3.8
django-environ==0.11.2
dj-database-url==2.1.0
django-debug-toolbar==4.2.0
whitenoise==6.6.0
django-redis==5.4.0
stripe==10.0.0
drf-spectacular==0.27.0

